requirements/dev.txt
-r base.txt

django-debug-toolbar>=4.3,<5.0
django-extensions>=3.2,<4.0
